Only run browser with active VNC connection
Browser based consoles have their own session timeout so it is
undesirable to start the console as soon as the container is started.

This change uses the -afteraccept and -gone arguments to x11vnc to start
and stop the selenium script and the browser.

Also -nevershare and -forever ensures that only one VNC session can be
active at once. Interaction with novncproxy may need to be revisited
with this.